+++ CMake/Options.cmake
@@ -442,7 +442,7 @@ carla_string_option (
 carla_string_option (
   CARLA_FASTDDS_VERSION
   "Target Fast-DDS version."
-  2.11.2
+  2.14.6
 )
 
 carla_string_option (
@@ -451,12 +451,40 @@ carla_string_option (
   v${CARLA_FASTDDS_VERSION}
 )
 
+# ==== CYCLONEDDS ====
+
+carla_string_option (
+  CARLA_CYCLONEDDS_VERSION
+  "Target CycloneDDS version."
+  0.10.5
+)
+
+carla_string_option (
+  CARLA_CYCLONEDDS_TAG
+  "Target CycloneDDS git tag."
+  ${CARLA_CYCLONEDDS_VERSION}
+)
+
+# ==== ZENOH-C ====
+
+carla_string_option (
+  CARLA_ZENOH_C_VERSION
+  "Target zenoh-c version."
+  1.8.0
+)
+
+carla_string_option (
+  CARLA_ZENOH_C_TAG
+  "Target zenoh-c release tag."
+  ${CARLA_ZENOH_C_VERSION}
+)
+
 # ==== FASTCDR ====
 
 carla_string_option (
   CARLA_FASTCDR_VERSION
   "Target Fast-CDR version."
-  1.1.x
+  2.2.x
 )
 
 carla_string_option (